Moved sipXconfig related changes from https://scm.sipfoundry.org/rep/*/main repositories: - 1068:1112 from  https://scm.sipfoundry.org/rep/sipXconfig/main - 596:598 from https://scm.sipfoundry.org/rep/sipXportLib/main - 578:579 from https://scm.sipfoundry.org/rep/sipXpbx/main
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main/sipXconfig@4835 a612230a-c5fa-0310-af8b-88eea846685b

+++ web/src/org/sipfoundry/sipxconfig/site/admin/AddUserRing.java
@@ -47,7 +47,7 @@ public abstract class AddUserRing extends BasePage implements PageRenderListener
 
         for (Iterator i = selections.getAllSelected().iterator(); i.hasNext();) {
             Integer userId = (Integer) i.next();
-            User user = getCoreContext().loadUser(userId.intValue());
+            User user = getCoreContext().loadUser(userId);
             callGroup.insertRing(user);
         }
